+++ skc_api/contract.go
@@ -1,4 +1,4 @@
-package contracts
+package skc_api
 
 type CardInfoResponse struct {
 	CardID        string `json:"cardID"`
